+++ app/Services/InvoiceServiceContract.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Modules\Invoices\Services;
+namespace App\Services;
 
 interface InvoiceServiceContract
 {
     public function getAllInvoices();
+    public function approveInvoice(string $id);
 }
